Moved CLI11 to externals and added license file
