+++ SU2_CFD/include/SU2_CFD.hpp
@@ -39,7 +39,7 @@
 #pragma once
 
 #include "../../Common/include/mpi_structure.hpp"
-#include "../../Common/include/toolboxes/CLI11.hpp"
+#include <CLI11.hpp>
 
 #include <ctime>
